[FIX] Deduplicate output-filename sanitizing helpers into pyhiv.split
