+++ src/pyhiv/split/__init__.py
@@ -3,5 +3,7 @@ from .split import (
     get_gene_region,
     get_present_gene_regions,
     map_ref_coords_to_alignment,
+    safe_output_label,
+    sequence_output_label,
     slugify_feature_name,
 )
